@@ -1,7 +1,8 @@
 {
   "name": "electron-react-typescript-webpack-boilerplate",
-  "version": "4.0.0",
+  "version": "4.1.0",
   "description": "Pre-configured boilerplate for Electron + React + TypeScript",
+  "type": "module",
   "main": "./dist/main.bundle.js",
   "scripts": {
     "start": "electron ./dist/main.bundle.js",
@@ -54,16 +55,16 @@
     "@types/jest": "^28.1.6",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.30.7",
-    "@typescript-eslint/parser": "^5.30.7",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
     "cross-env": "^7.0.3",
-    "electron": "^19.0.9",
-    "eslint": "^8.20.0",
+    "electron": "^20.0.1",
+    "eslint": "^8.21.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-import-resolver-webpack": "^0.13.2",
+    "eslint-import-resolver-typescript": "^3.4.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.6.0",
+    "eslint-plugin-jest": "^26.7.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-react": "^7.30.1",
     "eslint-plugin-react-hooks": "^4.6.0",
@@ -74,17 +75,17 @@
   "optionalDependencies": {
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
-    "electron-builder": "^23.1.0",
+    "electron-builder": "^23.3.3",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
-    "lodash": "^4.17.21",
     "rimraf": "^3.0.2",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.3.1",
-    "tsconfig-paths": "^4.0.0",
-    "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.10.0"
+    "tsconfig-paths": "^4.1.0",
+    "tsconfig-paths-webpack-plugin": "^4.0.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-merge": "^5.8.0"
   },
   "repository": {
     "type": "git",
