@@ -1,5 +1,5 @@
-import ipcAPI from '_preload/ipc-api';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { contextBridge } from 'electron';
+import ipcAPI from '_preload/ipc-api';
 
 contextBridge.exposeInMainWorld('ipcAPI', ipcAPI);
