@@ -4,9 +4,9 @@
 // Import the styles here to process them with webpack
 import '_public/style.css';
 
-import App from '_renderer/App';
 import * as React from 'react';
 import { createRoot } from 'react-dom/client';
+import App from '_renderer/App';
 
 const container = document.getElementById('app');
 // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
